Update testnet storybook config
